refactor(saves): sub-package __init__.py is re-export only
Move the SyncEngine class definition from sync_engine/__init__.py to
sync_engine/engine.py and the StatusService class definition from
status/__init__.py to status/service.py. Each sub-package __init__.py
now holds the package's contract-style docstring + a re-export of the
public class. Implementation lives in named modules.

Standard Python convention: __init__.py marks the package and exposes
the public API; primary code lives in named files. 600+ LOC class
definitions directly in __init__.py obscure the package's public
surface.

Top-level services/saves/__init__.py is unaffected — it stays as the
saves package's public API (SaveService façade).

Add the convention to CLAUDE.md so future sub-packages (slots/) follow
the pattern from the start.

Refs #257.
